SW_lab/sw_lab_part4/servizio_mail/requirements.txt
# Required
cherrypy == 18.6.0
paho-mqtt == 1.5.0
requests == 2.24.0
prompt-toolkit == 3.0.6
